introduced fastAPI context loader to remove tight coupling
