+++ config.yaml
@@ -27,7 +27,7 @@ llm:
   # Set to 'azure' or 'gemini' to choose the AI provider you previsouly set (above)
   provider: "gemini"
   system_prompt: "You are a helpful assistant that answers questions based on the provided documentation. Reply in the same language as the question, unless otherwise specified."
-  default_k: 3
+  default_k: 10
   # Azure OpenAI LLM
   azure:
     temperature: 0.1
@@ -44,7 +44,7 @@ llm:
 # Document Processing Configuration
 document_loader:
   chunk_size: 300
-  chunk_overlap: 15
+  chunk_overlap: 30
   docs_directory: "docs"
   # directory of the pdf files that are scanned (images) and not proper text
   ocr_docs_dir: "ocr_docs"
