+++ src/ai/vector_store_service/__init__.py
@@ -1,3 +1,3 @@
-from .vestor_store_facade import VectorStoreFacade
+from .vector_store_facade import VectorStoreFacade
 
 __all__ = ["VectorStoreFacade"]
