+++ src/ai/vector_store_service/faiss/faiss_vector_store_initializer.py
@@ -1,18 +1,14 @@
-from src.ai.vector_store_service.faiss_vector_store import FaissVectorStore
-from src.ai.document_loaders.document_loader_facade import DocumentLoaderFacade
 import logging
-from typing import List, Dict
+from src.ai.vector_store_service.faiss.faiss_vector_store import FaissVectorStore
+from src.ai.document_loaders.document_loader_facade import DocumentLoaderFacade
+from src.common.app_context import get_app_context
 
-class VectorStoreFacade:
+class VectorStoreInitializer:
+    """Handles loading or building the vector store."""
+
     def __init__(self):
-        self.__vector_store = FaissVectorStore()
+        self.__vector_store = FaissVectorStore(get_app_context().state.embedder)
 
-    def add_documents(self, texts: List[str], metadatas: List[Dict]):
-        self.__vector_store.add_documents(texts, metadatas)
-
-    def search(self, question: str, k: int) -> List[Dict]:
-        return self.__vector_store.search(question, k=k)
-    
     def initialize_vector_store(self):
         """
         Initialize the vector store from documents.
@@ -24,7 +20,6 @@ class VectorStoreFacade:
             logging.info("Nothing found in vector store, building from documents")
             
             # Use facade to load all documents from all sources
-            # No LLM needed - facade creates it internally if needed!
             document_loader_facade = DocumentLoaderFacade()
             texts, metadatas = document_loader_facade.load_all_documents()
 
@@ -37,4 +32,3 @@ class VectorStoreFacade:
             logging.info("Vector store initialized successfully with documents") 
         else:
             logging.info("Vector store loaded from disk")
-
